+++ changelog.md
@@ -1,6 +1,7 @@
 # 0.6.5
 - Minor fix in the setFilterValue value (base-db.component) - gte/gt & lte/lt now work correctly.
 - Upgraded the migrations module's insertStaticData functionality - it now inserts static data according to the seeding order first, then the rest of the data.
+- Added the "objSort" method to the toolbelt.
 
 # 0.6.4
 - Fixed to stupid errors caused by rushed release.